ported to next.js, backend api integration
integrated with backend api (app.py) at http://127.0.0.1:5000/api/sites.
automatically populates the grid with site data from the endpoint. still
using iframes but they are now wrapped in a div that directs to the
website on click. removed scrolling in the iframe so that it functions
as a simple preview